VDP: Attempt to fix V-scroll
Apparently V-scroll isn't aligned with the screen: it actually
scrolls with the plane. This results in a bug on real hardware when
the H-scroll value isn't a multiple of 16, as the screen will be
showing 21 vertical columns, when only 20 V-scroll values exist in
VSRAM. This is noted in section 4 of the Genesis Software
Developement Manual (Complement), 'PRECAUTIONS WHEN THE CONTROL A AND
B SCREENS ARE USED'.