Support H-Int intervals and overhaul Int timing
I am extremely unfamiliar with NTSC and PAL, so previously I was
assuming that the display would output batches of 224 (or 240)
scanlines one after another. However, it turns out that the Mega
Drive actually outputs at a lower vertical resolution than NTSC and
PAL TVs support, and it merely displays nothing in the spare
scanlines to fill the space. This affects the timing of H-Int and
V-Int. Luckily for me, page 7 of the Genesis Software Manual
developer document explains this.

Also, now H-Int intervals are supported, fixing the water effect in
Labyrinth Zone in Sonic 1, and Chemical Plant Zone and Aquatic Ruin
Zone in Sonic 2.