Fix integer casting/promotion stuff
It's tricky to test that code supports platforms where `int` is
16-bit.

That said, brownbot.mooo.com gives you a somewhat-convenient way to
compile code for the 68k, which can be configured with the `-mshort`
flag to have a 16-bit `int` type. Combined with the
`-Wshift-overflow=2` flag, this is a handy way to detect overflow and
other improper use of the `int` type.