Remove the internal ROM buffer
Now, the emulator accesses the ROM through frontend callbacks.

By accessing the ROM through callbacks instead of direct array
accesses, it is now theoretically possible for the ROM to be streamed
from disk in chunks instead of having the whole file loaded into
memory at once. This is a portability improvement, as it would
benefit platforms with extremely limited memory, especially with
particularly-large ROMs.

I believe that this gives clownmdemu an edge over Genesis Plus GX,
which appears to work around the issue by limiting the size of its
internal ROM buffer on memory-limited platforms instead of using a
solution like this one.