Split persistent/configuration data from state
Now the there is state data, configuration data, and persistent data.

State data is state (obviously), which is altered by clownmdemu.

Configuration data is modifiable by the frontend, but not by
clownmdemu.

Persistent data is data is not modifiable by the frontend, nor by
clownmdemu: the only time that persistent data is modified is when it
is first created during initialisation. It is intended for lookup
tables and other constant precalculated data.

By splitting configuration data and persistent data from state, their
redundant duplication by the rewind buffer is avoided, as is the
pointless storage of this data in save states. Additionally, this
prevents rewinding and loading save states from undoing frontend
configuration such as changing region.