Document the interlace stuff some more
With clownmdemu's current API, I think interlace mode 1 can actually
be implemented! It just won't be noticable with the reference
frontend since it combines the scanlines into a flat bitmap.

This emulator isn't intended to be super low level or anything, but I
do think this can be emulated a bit more: interlace mode 1 basically
produces a 320x480 image which is just a 320x240 image with doubled
lines. If my emulator were to output a native 320x480 (perhaps
stretched to 640x480) image, and then

For *absolute* authenticity, one
could theoretically have this emulator output directly to an old TV.
If