M68k: Generate a better interpreter.
Now, instead of each step of execution being its own switch
statement, all steps are part of a single switch statement. To avoid
making the programmer navigate a ton of duplicate code, the optimised
interpreter is generated from a modified version of the original
interpreter.

This makes the 68k interpreter around 2 to 3 times faster.