Add support for second way of detecting PAL mode.
There are two ways to detect PAL mode on a Mega Drive: the IO area at
0xA10001, and by reading the VDP's control port.

Previously, this emulator only supported the former, but now it also
supports that latter. This is required for Sonic the Hedgehog 2,
which relies on the latter method for deciding the timing of its SMPS
(AKA 'Sound Source' & 'Sound Sorce') sound driver.