Add support for the YM2612's timers.
Dammit: I spent so much time refactoring `FM_Update` to operate on
cycles instead of audio frames, but it turns out that the timers
measure in exact multiples of audio frames, making all that work
redundant!

Still, I suppose using cycles instead of frames is the proper thing
to do.

With this commit, the YM2612's Timer A and Timer B are supported.
This enables games such as Vectorman and Castlevania Bloodlines to
finally output sound. It also gets my 'Sonic 2 except the music goes
as fast as you do' ROM hack to output audio as well.