Convert VRAM to 8-bit instead of 16-bit.
According to the Genesis Software Manual, CRAM and VSRAM are 16-bit,
but VRAM is 8-bit. This affects how the DMA transfers work, as it
enables odd addresses to be used. It also means that the value used
to fill VRAM is a byte and not a word.

This fixes the graphics in Mega Man: The Wily Wars.