PSG: Fix frequencies being slightly off.
I should have taken SMSPower more seriously, because it appears to be
totally correct: it really does seem to 'tick' when it hits zero, not
when it tries to go below it.

https://www.smspower.org/Development/SN76489

This fixes the PSG always sound slightly off-key. I'm a little
tone-deaf, so I couldn't tell that this was the case without some
extreme examples, like Sonic 1's Special Stage theme.