Only the 68k is reset by a soft reset.
My custom DAC driver overwrites its initialisation code with a mixing
buffer. Normally this is okay because the Z80 is never reset, but
this mistake was doing just that. As a result, the mixing buffer was
being interpreted as code and causing the frontend's log window to be
flooded with errors.

As documented in various places on the internet, the soft-reset
signal ("VRES") only resets the Motorola 68000:
https://www.retrosix.wiki/mega-drive-circuit-sega-mega-tech